+++ HISTORY.rst
@@ -9,6 +9,8 @@ Unreleased
 Changes
 ~~~~~~~~~~~~~~~~~~~~~
 * add additional metrics (Top-1, Top-5, MAE, MSE) in process `model-tester` execution results (GEOIM-163)
+* add fetching of taxonomy definition during `batch-creation` process execution from input URL (GEOIM-161)
+* save retrieved taxonomy definition to generated dataset batch (GEOIM-162)
 
 1.2.0 (2019-06-11)
 ---------------------